refactor: User data class 별도의 파일로 분리
