+++ app/src/main/java/com/beside153/peopleinside/model/common/User.kt
@@ -1,0 +1,12 @@
+package com.beside153.peopleinside.model.common
+
+data class User(
+    val userId: Int,
+    val jwtToken: String,
+    val nickname: String,
+    val mbti: String,
+    val birth: String,
+    val gender: String,
+    val email: String? = "",
+    val isMember: Boolean
+)
