+++ app/src/main/java/com/beside153/peopleinside/viewmodel/community/PostDetailViewModel.kt
@@ -5,9 +5,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.beside153.peopleinside.base.BaseViewModel
 import com.beside153.peopleinside.model.common.CreateContentRequest
+import com.beside153.peopleinside.model.common.User
 import com.beside153.peopleinside.model.community.comment.CommunityCommentModel
 import com.beside153.peopleinside.model.community.post.CommunityPostModel
-import com.beside153.peopleinside.repository.User
 import com.beside153.peopleinside.repository.UserRepository
 import com.beside153.peopleinside.service.community.CommunityCommentService
 import com.beside153.peopleinside.service.community.CommunityPostService
