+++ app/src/main/java/com/beside153/peopleinside/viewmodel/mypage/editprofile/EditProfileViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.viewModelScope
 import com.beside153.peopleinside.App
 import com.beside153.peopleinside.base.BaseViewModel
 import com.beside153.peopleinside.common.exception.ApiException
+import com.beside153.peopleinside.model.common.User
 import com.beside153.peopleinside.model.editprofile.EdittedUserInfo
-import com.beside153.peopleinside.repository.User
 import com.beside153.peopleinside.repository.UserRepository
 import com.beside153.peopleinside.service.UserService
 import com.beside153.peopleinside.util.Event
