+++ app/src/main/java/com/beside153/peopleinside/viewmodel/mypage/setting/DeleteAccountViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.beside153.peopleinside.base.BaseViewModel
+import com.beside153.peopleinside.model.common.User
 import com.beside153.peopleinside.model.user.ResonIdModel
 import com.beside153.peopleinside.model.withdrawal.WithDrawalReasonModel
-import com.beside153.peopleinside.repository.User
 import com.beside153.peopleinside.repository.UserRepository
 import com.beside153.peopleinside.service.UserService
 import com.beside153.peopleinside.service.WithDrawalService
