+++ app/src/main/java/com/beside153/peopleinside/viewmodel/onboarding/ContentChoiceViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.beside153.peopleinside.base.BaseViewModel
+import com.beside153.peopleinside.model.common.User
 import com.beside153.peopleinside.model.mediacontent.OnBoardingChosenContentModel
 import com.beside153.peopleinside.model.mediacontent.OnBoardingContentModel
-import com.beside153.peopleinside.repository.User
 import com.beside153.peopleinside.repository.UserRepository
 import com.beside153.peopleinside.service.UserService
 import com.beside153.peopleinside.service.mediacontent.MediaContentService
