+++ app/src/main/java/com/beside153/peopleinside/viewmodel/onboarding/signup/SignUpUserInfoViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.beside153.peopleinside.base.BaseViewModel
 import com.beside153.peopleinside.common.exception.ApiException
-import com.beside153.peopleinside.repository.User
+import com.beside153.peopleinside.model.common.User
 import com.beside153.peopleinside.repository.UserRepository
 import com.beside153.peopleinside.service.UserService
 import com.beside153.peopleinside.util.Event
